backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
python-multipart
passlib[bcrypt]
PyJWT
pydantic
pandas
openpyxl
